Menambahkan API buat produk, Kategori, Cart, Transaction

+++ database/migrations/2025_05_04_181822_create_transactions_table.php
@@ -9,7 +9,7 @@ return new class extends Migration {
         Schema::create('transactions', function (Blueprint $table) {
             $table->id();
             $table->foreignId('user_id')->constrained()->onDelete('cascade');
-            $table->decimal('total_price', 10, 2);
+            $table->integer('total_price');
             $table->enum('status', ['pending', 'paid', 'cancelled'])->default('pending');
             $table->timestamps();
         });
